Rename targets.mutate component to discovery.relabel
