+++ component/discovery/relabel/relabel_test.go
@@ -1,11 +1,11 @@
-package mutate_test
+package relabel_test
 
 import (
 	"testing"
 	"time"
 
 	"github.com/grafana/agent/component/discovery"
-	"github.com/grafana/agent/component/targets/mutate"
+	"github.com/grafana/agent/component/discovery/relabel"
 	"github.com/grafana/agent/pkg/flow/componenttest"
 	"github.com/grafana/agent/pkg/river/parser"
 	"github.com/grafana/agent/pkg/river/vm"
@@ -15,8 +15,8 @@ import (
 func TestRelabelConfigApplication(t *testing.T) {
 	riverArguments := `
 targets = [ 
-	{ "__meta_foo" = "foo", "__meta_bar" = "bar", "__address__" = "localhost", "instance" = "one", "app" = "backend", __tmp_a = "tmp" },
-	{ "__meta_foo" = "foo", "__meta_bar" = "bar", "__address__" = "localhost", "instance" = "two", "app" = "db", "__tmp_b" = "tmp" },
+	{ "__meta_foo" = "foo", "__meta_bar" = "bar", "__address__" = "localhost", "instance" = "one",   "app" = "backend",  "__tmp_a" = "tmp" },
+	{ "__meta_foo" = "foo", "__meta_bar" = "bar", "__address__" = "localhost", "instance" = "two",   "app" = "db",       "__tmp_b" = "tmp" },
 	{ "__meta_baz" = "baz", "__meta_qux" = "qux", "__address__" = "localhost", "instance" = "three", "app" = "frontend", "__tmp_c" = "tmp" },
 ]
 
@@ -55,7 +55,7 @@ relabel_config {
 	regex  = "__meta(.*)|__tmp(.*)|instance"
 }
 `
-	expectedExports := mutate.Exports{
+	expectedExports := relabel.Exports{
 		Output: []discovery.Target{
 			map[string]string{"__address__": "localhost", "app": "backend", "destination": "localhost/one", "meta_bar": "bar", "meta_foo": "foo", "name": "one"},
 		},
@@ -64,11 +64,11 @@ relabel_config {
 	file, err := parser.ParseFile("agent-config.river", []byte(riverArguments))
 	require.NoError(t, err)
 
-	var args mutate.Arguments
+	var args relabel.Arguments
 	err = vm.New(file).Evaluate(nil, &args)
 	require.NoError(t, err)
 
-	tc, err := componenttest.NewControllerFromID(nil, "targets.mutate")
+	tc, err := componenttest.NewControllerFromID(nil, "discovery.relabel")
 	require.NoError(t, err)
 	go func() {
 		err = tc.Run(componenttest.TestContext(t), args)
